feature-home: extract AddonPillow into its own file
The add-on pillow composable and its flow-to-pillow mapping lived in the middle
of the large HomeDestination file. Move both into AddonPillow.kt and have it
take a FlowType directly, so HomeDestination no longer carries the PillowType
mapping or its imports.
